Change contrastive loss
